% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapyranhia-package.R
\docType{package}
\name{datapyranhia-package}
\alias{datapyranhia}
\alias{datapyranhia-package}
\title{datapyranhia: Jeu de données pour les formations Pyranhia}
\description{
Ce package fournit des jeux de données utilisés dans les formations Pyranhia, facilitant l'apprentissage et la manipulation de données en R.
}
\author{
\strong{Maintainer}: Thelma Panaïotis \email{thelmapana@proton.me}

}
\keyword{internal}
