% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing.R
\docType{data}
\name{housing}
\alias{housing}
\title{Données California Housing}
\format{
Un tibble avec 20640 lignes et 10 variables :
\describe{
\item{longitude}{dbl Longitude du groupe de blocs}
\item{latitude}{dbl Latitude du groupe de blocs}
\item{housing_median_age}{dbl Âge médian des logements}
\item{total_rooms}{dbl Nombre total de pièces}
\item{total_bedrooms}{dbl Nombre total de chambres}
\item{population}{dbl Nombre total de personnes vivant dans la zone}
\item{households}{dbl Nombre de ménages dans la zone}
\item{median_income}{dbl Revenu médian des ménages}
\item{median_house_value}{dbl Valeur médiane des logements}
\item{ocean_proximity}{chr Proximité à l’océan
(\code{NEAR BAY}, \code{<1H OCEAN}, \code{INLAND},
\code{NEAR OCEAN}, \code{ISLAND})}
}
}
\source{
California Housing dataset publié sous licence CC0 (domaine public) :
\url{https://squirrel-datasets-core.readthedocs.io/en/stable/dataset_links/california_housing.html}

Les données originales proviennent du recensement américain de 1990 et sont
mentionnées dans Pace & Barry (1997), \emph{Sparse spatial autoregressions}.
}
\usage{
housing
}
\description{
Jeu de données sur les logements en Californie issu du recensement américain
de 1990, couramment utilisé pour des exemples de régression. Chaque ligne
représente un groupe de blocs de recensement (\emph{census block group}).
}
\details{
Chaque observation correspond à un groupe de blocs de recensement
(\emph{census block group}) du recensement américain de 1990. Les variables
sont agrégées à ce niveau géographique.
}
\keyword{datasets}
