% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sportifs.R
\docType{data}
\name{sportifs}
\alias{sportifs}
\title{Données de performance des sportifs}
\format{
Un tibble avec 500 lignes et 4 variables :
\describe{
\item{heures}{dbl Heures d’entraînement par semaine}
\item{experience}{dbl Annés d’expérience}
\item{score}{dbl Score de performance}
\item{set}{chr Dataset split : \code{"Train"} ou \code{"Test"}}
}
}
\usage{
sportifs
}
\description{
Jeu de données synthétiques pour illustrer les principaux concepts du machine learning.
}
\keyword{datasets}
