% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sportifs_cart.R
\docType{data}
\name{sportifs_cart}
\alias{sportifs_cart}
\title{Données de performance des sportifs adapté pour les CART.}
\format{
Un tibble avec 500 lignes et 4 variables :
\describe{
\item{heures}{dbl Heures d’entraînement par semaine}
\item{experience}{dbl Annés d’expérience}
\item{score}{dbl Score de performance}
\item{set}{chr Dataset split : \code{"Train"} ou \code{"Test"}}
}
}
\usage{
sportifs_cart
}
\description{
Jeu de données synthétiques pour illustrer les concepts des arbres de régression et de classification (CART).
}
\keyword{datasets}
