% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine.R
\docType{data}
\name{wine}
\alias{wine}
\title{Données de qualité du vin}
\format{
Un tibble avec 1143 lignes et 12 variables :
\describe{
\item{fixed_acidity}{dbl Acidité fixe}
\item{volatile_acidity}{dbl Acidité volatile}
\item{citric_acid}{dbl Acide citrique}
\item{residual_sugar}{dbl Sucre résiduel}
\item{chlorides}{dbl Chlorures}
\item{free_sulfur_dioxide}{dbl Dioxyde de soufre libre}
\item{total_sulfur_dioxide}{dbl Dioxyde de soufre total}
\item{density}{dbl Densité}
\item{pH}{dbl Niveau de pH}
\item{sulphates}{dbl Sulfates}
\item{alcohol}{dbl Teneur en alcool}
\item{quality}{int Score de qualité sensorielle compris entre 0 et 10}
}
}
\source{
Cortez, P., Cerdeira, A., Almeida, F., Matos, T., & Reis, J. (2009).
\emph{Modeling wine preferences by data mining from physicochemical properties}.
Decision Support Systems, 47(4), 547–553.

Données disponibles sur le UCI Machine Learning Repository :
\url{https://archive.ics.uci.edu/dataset/186/wine+quality}.
Licence Creative Commons Attribution 4.0 International (CC BY 4.0).
}
\usage{
wine
}
\description{
Ce jeu de données contient des mesures physico-chimiques de vins ainsi qu’un
score de qualité sensorielle. Il est couramment utilisé pour des analyses de
régression et de classification visant à relier les propriétés chimiques du
vin à la qualité perçue.
}
\keyword{datasets}
